use -execdir option from find instead of xargs
This is so that if there is more than one file that does not
validate they will all be reported in the error output.
Invocation with '\;' reports all errors, but does not alter exit
status of find, so is no good as a check.
Invocation with '+' stops with exit code one on first error.

Change-Id: I9318c81b78d0d6b45db5abcd32f69522e6197ece