Change to use tempest run for whitelist test
ostestr with whitelist is not working in the latest ostestr code.
The reason is os-testr will start to use "stestr"
instead of old "testrepository" since late July.
And it is not working when specifying whitelist file option.

Other ostestr test with -n option is still working.
The command could be changed back after "stestr" bug fix.

Change-Id: I147b9cf3192a50facc4584e964690406a45c3c42