Add librarian path to jenkins job
A better solution will probably be to add it to one of the many rc files
that are sourced on user login. However, jenkins jobs are run by
/bin/sh, and is not a login shell, so we can't use .profile or .bashrc
to do what we want.

Editing the system level rc files might cause side effects, so we just
go along with this for now.

Change-Id: I0943609f8c299aa48be852b2fff7ca12d76f5d39