Disable scanning of 'Build Queue' for blocking jobs
Currently we are having succesive failures in tempest nagios jobs, with
the first job being a timeout and the last few failures being
insufficient quota.

This is due to cleanup jobs are being blocked when there is a job in the
queue. This normally happens when a tempest job fails and immediately
queues another job to run. The queued job will actually prevent cleanup
from running, causing a string of continuous failures. After 5 failures
the retying mechanism gives up, and cleanup job manages to run.

Fix this by removing the option to scan build queues and just scan
running jobs [1].

[1]: https://docs.openstack.org/infra/jenkins-job-builder/properties.html#properties.build-blocker

Change-Id: If874f3b75ed728ad77268afdd70e5a7c2cd7fd78