Add default GERRIT_PROJECT in site-deploy job
The site-deploy job depends on GERRIT_PROJECT being set,
which only happens when the job is triggered by a Gerrit
event (e.g., change-merged). To trigger the job manually
(or as a post-build step of another job) we need to ensure
there's a default value set. This value never actually
changes since each job is templated to be specific to a
project.

Change-Id: I8b96ea42a472ae961ceffae5e8e1a1474b438371