Also deploy r10k environment on comment change
This is because the r10k-cleanup job assumes each change generates a
r10k environment (and that there is a branch named after the commit sha
in the r10k repo). If we skip commits were only comment was changed, the
cleanup script errors out due to the missing branch in r10k repo.

Change-Id: Ica270671a92c37fdb46adaa9639a6079cd77a83a