Add blocker to make sure deploy runs after cleanup
This job must be run AFTER *-r10k-cleanup finishes, as *-r10k-cleanup
will push the merged commmit to the master branch in r10k repo which
this job picks up from.

In Change I2f0caa400a58de24862e6fff850198f7246b619f we added a quiet
period to it so, but this is insufficient as there are cleanup job that
runs for 14 secs.

Add a blocker to make sure cleanup is really done before the deploy job
runs.

Change-Id: Icdb0d8f714a6da3d2b326c40c846f70150a9b373