Fix container build jobs triggering test cloud upgrade
The 'publisher' method does not work with pipelines - 'publisher'
translates to 'Post-build Actions' that doesn't show up in Jenkins Web
UI when you are configuring a 'Pipeline' type job. It shows up in a
'Freestyle' type job.

Use the 'post' in Pipelines as documented[1]. Also, add a parameter
'GERRIT_PROJECT' to the deploy job as Jenkins doesn't fill that in if
triggered via a pipeline job.

[1]: https://www.jenkins.io/doc/book/pipeline/syntax/#post

Change-Id: I52a00b0bb989df8f916bd8928837eac8f165e58d