Refactor -cleanup and -deploy-control-repo jobs
When a puppet-site-xxx control repo change is merged, two jobs normally
run.

cleanup: cleans up the branches created for each change, and pushes
the merged commit (now on HEAD of master) to a r10k repo.

deploy-control-repo: deploys the change from the r10k repo. We deploy
from a r10k repo and not the original repo because there are special
branches created for r10k to work, and we don't want these branches in
the original repo.

This change moves the pushing of merged commit from the 'cleanup' job to
the 'deploy-control-repo' job.

This is because the 'deploy-control-repo' job is now called by other
jobs, notably the container building jobs to deploy a new change to
rctest repo. For that to work, we have to push the merged commit to the
r10k repo.

Change-Id: I246092cb75a18814be1385dc2106b469ab73274e