Change "{name}-puppet-unit" template to a variable based job name
This change will allow us to create new jobs (with different names,
based on name-suffix)
based on the same template, with the same base settings as in the
template, overriding/adding the ones we want.

sample usage:

- project:
    name: puppet-jacob-rocks
    jobs:
      - '{name}-puppet-unit':
         organisation: 'internal'

will create a job named "puppet-jacob-rocks-puppet-unit"

while

- project:
    name: puppet-jacob-rocks
    name-suffix: -puppet7
    jobs:
      - '{name}-puppet-unit':
         organisation: 'internal'

will create a job named "puppet-jacob-rocks-puppet-unit-puppet7"

This change will not affect current jobs, just add the ability to have
new ones with a name-suffix.

based on:
https://jenkins-job-builder.readthedocs.io/en/latest/definition.html

Change-Id: I61549f16c84b0c43283349a085a0e2ed75c9f207