Mimic upstream testing for octocatalog-diff
In the current version of octocatalog-diff, `bundler exec rake test`
will not be able to run on a normal focal box due to older ciphers being
used in the tests.

To get around this, upstream now uses a docker container for testing, which they
have an edited openssl.cnf.

We mimic upstream testing so that this can run on our focal boxes.
Copied workflow from `.github/workflows/main.yml` and `Dockerfile`

Change-Id: Id3582704b8a92f6ef547db5dfd5037f1641e2c3b