Fix for problem of Helm Chart update ordering
The problem is that if multiple commits are merged at the same time
and the builds run in parallel, they may *complete* or of order.
This can result in the latest Chart version not containing the
latest commit.  The fix is to set 'concurrent: false' for the affected
container builds.

Change-Id: Ic4df344c387481a2e86cc33ba280bacef14f920f