Add internal git secrets repo
internal/git-secrets-puppet-nectar will hold the sensitive hiera keys in
internal and third party repos

The `git clean -ffdx` is needed because jenkins-job-builder cannot be set to
'clean untracked nested repos' in the scm - git module. That option does
not exists in jjb but exists in Jenkins.

Added ssh-agent credential as it is needed for git clone via SSH

Change-Id: I5d2f62130c3626a07bfbecb277f9ecba7f74f51c