Add kolla_profile_name for kolla build job
In some cases, the kolla profile name is not same as the project name,
so we can override kolla profile name.

For ec2-api service, the project name/repo name is ec2-api, but the user
in kolla image is called ec2api. So we have to override the
kolla_profile_name to ec2api.

Change-Id: I83899fa409260f3333edd1c7767c0ba9c7d8be49