Add auto configure tox job template
Add a job template '{name}-os-tox-python3' that mimics openstack
upstream jobs[1]. This allow us to just specify one job for an openstack
project, and it will figure out the appropriate ubuntu distro / python
version to use for tox tests. This saves each operator having to update
python version for their service when going to Jammy/Zed.

Upstream jobs currently run the following:

- Yoga (openstack-python3-yoga-jobs)
  - pep8 on Focal
  - py36 on Bionic
  - py39 on Focal
- Zed (openstack-python3-zed-jobs)
  - pep8 on Focal
  - py38 on Focal
  - py39 on Focal
  - py310 on Jammy (non-voting)

We customise our jobs do to the following:

- Yoga
  - pep8 on Focal
  - py38 on Focal
  - py310 on Jammy
- Zed
  - pep8 on Focal
  - py310 on Jammy

This is because we support Yoga on both Focal and Jammy, and we support
Zed on Jammy.

[1]: https://docs.openstack.org/infra/openstack-zuul-jobs/project-templates.html

Change-Id: Iac1e38ae3a76a81b1596198e7a1e268fce79e0e4