Create separate jobs for midonet testing
The site and site midonet jobs use different projects but the
same tempest user. So for same site, site cleanup job should be blocked for both site and site midonet jobs. (Because like keypair which is shared by the user will be deleted by the cleanup job, and this will lead to the failure of ongoing testing). And same for site midonet cleanup job.

For example, 'tempest-cleanup-site-melbourne-qh2-midonet' should be blocked when 'tempest-nagios-site-melbourne-qh2-coe' is running. This is because ospurge purges keypairs which belongs to the user, not project.

We create separate jobs for midonet testing and its cleanup so that we can block the relevant jobs easily.


Change-Id: Ie5bacd62675329b1c387c62eff7da32a7fe716c1