Hardcode the source/target host for compute adv job in rctest
We want to run the periodical compute adv tests in rctest. One reason it
is not always working is during the upgrade, migration from newer node
to older one will fail as expected. So we hardcode the source/target host
in this job, as we usually upgrade the compute nodes in the order cc3->cc4->cc5

Change-Id: I1a24587f77bd5378ed85e38ddc1f934877c25bf2